# doc/api.yaml
openapi: 3.0.0
info: 
  description: WASAText exam project Rest API
  title: WASAText
  version: 1.0.0
servers:
- url: http://localhost:3000
components:
  schemas:
...
    chat: 
      title: chat
      description: the chat between two users or more users
      type: object
      properties:
        c_id:
          description: unique identifier
          type: integer
        c_userIds:
          description: the id of the user of the chat
          type: array
          minItems: 1
          maxItems: 32
          items:
            type: integer
            minimum: 0
        c_textMessageList:
          description: the ids of the messages sent in this conversation
          type: array
          minItems: 0
          maxItems: 1024
          items:
            type: integer
            minimum: 0
        c_groupName:
          description: the name of the group
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 256
        c_groupPhoto:
          description: the photo of the group
          type: string
          format: binary
          minLength: 1
          maxLength: 1024
        c_description:
          description: the description of the group
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 256
        c_group:
          description: if the chat is a group chat or not
          type: boolean
  securitySchemes:
    BearerAuth: 
      type: http 
      scheme: bearer 
      bearerFormat: JWT   
...
paths:
  /session:
    post:
      tags: ["user profile"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              description: temp
              type: object
              properties:
                l_name:
                  description: temp
                  type: string
                  pattern: 'ˆ.*?$'
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                description: temp
                type: object
                properties:
                  l_identifier:
                    description: id dell' utente
                    type: string
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 32
...
  /users/{id}/conversations:
    get:
      tags: [chat management]
      summary: gets all the conversationof the user
      description: |
         gets all the conversation of the user 
      operationId: getMyConversations
      security:
        - BearerAuth: []   
      parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      responses:
        '200':
          description: list of the users chat
          content:
            application/json:
              schema:
                description: temp
                type: array
                minItems: 0
                maxItems: 32
                items:
                  $ref: '#/components/schemas/chat'
        '401':
          description: not logged in
        '404': 
          description: The requested resource could not be found but may be available in the future
...
  /users/{id}/conversations/{conversationId}/messages:
    post:
      tags: ["message management"]
      summary: send a message in a chat.
      description: |
        send a message to a chat or a group
      operationId: sendMessage
      security:
        - BearerAuth: []   
      parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      - name: conversationId
        in: path
        description: id conversazione
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      requestBody: 
        description: the messages specifications
        content:
          multipart/form-data:
            schema:
              description: temp
              type: object
              properties:
                messageText:
                  description: the text part of the message
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
                photoText:
                  description: the photo part of the message
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
        required: true
      responses:
        '204': 
          description: message successfully sent
        '401':
          description: not logged in
  /users/{id}/conversations/{conversationId}/messages/{messageId}:
    parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0 
      - name: conversationId
        in: path
        description: id conversation
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0 
      - name: messageId
        in: path
        description: id messaggio
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0 
    post:
      tags: ["message management"]
      summary: send a text from a chat to another
      description: send a text from a chat to another
      operationId: forwardMessage
      security:
        - BearerAuth: []   
      requestBody: 
        description: new chat id
        content:
          application/json:
            schema:
              description: the id of the chat where the message is going to be forwarded
              type: object
              properties:
                chatId:
                  description: temp
                  type: integer
                  minimum: 0
      responses:
        '204':
          description: message forwarded
        '401':
          description: not logged in
        '406': 
          description: invalid text
    delete:
      tags: ["message management"]
      summary: deletes a message
      description: delete a message previusly sent
      operationId: deleteMessage
      security:
        - BearerAuth: []   
      responses:
        '204': 
          description: message successfully deleted
        '401':
          description: not logged in
        '403':
          description: unauthorized to delete this message
        '404':
          description: The requested resource could not be found 
  /users/{id}/conversations/{conversationId}/messages/{messageId}/comments:
    parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      - name: conversationId
        in: path
        description: id conversation
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      - name: messageId
        in: path
        description: id message
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
    post:
      tags: ["message management"]
      summary: comment another message
      description: comment another message 
      operationId: commentMessage
      security:
        - BearerAuth: []   
      requestBody: 
        description: message to comment
        content:
          multipart/form-data:
            schema:
              description: temp
              type: object
              properties:
                messageText:
                  description: the text part of the message
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
                photoText:
                  description: the photo part of the message
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
        required: true
      responses:
        '204': 
          description: message commented
        '401':
          description: not logged in
        '406': 
          description: invalid text
    delete:
      tags: ["message management"]
      summary: uncomment another message
      description: uncomment another message
      operationId: uncommentMessage
      security:
        - BearerAuth: []   
      responses:
        '204': 
          description: comment successfully uncommented
        '401':
          description: not logged in
        '403':
          description: unauthorized to delete this message
        '404':
          description: The requested resource could not be found 
  /users/{id}/groups:
    post:
      tags: ["chat management"]
      summary: add a user to a group
      description: add a user to a group. If successful returns the group
      operationId: addToGroup
      security:
        - BearerAuth: []   
      parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      requestBody: 
        description: the user to add and the group where he is going to be added
        content:
          application/json:
            schema:
              description: temp
              type: object
              properties:
                adg_utente_id:
                  description: the id of the user to add to the group
                  type: integer
                adg_group_id:
                  description: the id of the group where the user is going to be added
                  type: integer
      responses:
        '204': 
          description: user added to the group
        '401':
          description: not logged in
        '403':
          description: not authorized
        '406': 
          description: invalid user
...
  /users/{id}/conversations/{conversationId}/groupName:
    post:
      tags: ["chat management"]
      summary: change the name of a group
      description: change the name of a group . If successful returns the group id
      operationId: setGroupName
      security:
        - BearerAuth: []   
      parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      - name: conversationId
        in: path
        description: id conversation
        required: true
        schema:
          description: temp
          type: object
          properties:
            sgn_con_id:
              description: temp
              type: integer
      requestBody:
        description: the name of the group
        content:
          application/json:
            schema:
              description: temp
              type: object
              properties:
                sgn_name:
                  description: the new name of the group
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
        required: true
      responses:
        '204':
          description: group name changed
        '401':
          description: not logged in
        '403':
          description: not authorized
...